requirements.txt
# AI CAPTCHA Solver - Dependencies
# Core ML and Data Processing
tensorflow>=2.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0

# Image Processing and CAPTCHA Generation
captcha>=0.4.0
Pillow>=10.0.0

# OCR Baseline Comparison
pytesseract>=0.3.10

# Web Interface and Utilities
streamlit>=1.28.0
tabulate>=0.9.0

# Optional: For Google Colab training (automatically available in Colab)
# google-colab (only needed if running locally with Colab features)
